+++ userprofile/views.py
@@ -39,7 +39,7 @@ class UserDetailView(LoginRequiredMixin, LoggedInPermissionMixin, TemplateView):
         ) + get_active_plugins(plugin_type='site_app')
         for plugin in plugins:
             p_settings = app_settings.get_setting_defs(
-                plugin, APP_SETTING_SCOPE_USER, user_modifiable=True
+                APP_SETTING_SCOPE_USER, plugin=plugin, user_modifiable=True
             )
             for s_key, s_val in p_settings.items():
                 yield {